requirements.txt
torch --index-url https://download.pytorch.org/whl/cu128
torch_geometric
torch_scatter
pytorch_lightning
jupyter
scipy
pandas
matplotlib